+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.1.2
+-----
+- Add offline installment feature
+- Fix configuration page JS issue
+- Default Installment setting disabled
+
 2.1.0
 -----
 Add All product installment feature